+++ BxInputController/Sources/Rows/Boolean/View/BxInputCheckRowBinder.swift
@@ -15,7 +15,7 @@ import UIKit
 
 /// Binder for check box Row subclasses
 open class BxInputCheckRowBinder<Row : BxInputCheckRow, Cell>: BxInputStandartTextRowBinder<Row, Cell>
-where Cell : UITableViewCell, Cell : BxInputStandartTextCellProtocol
+where Cell : UITableViewCell, Cell : BxInputFieldCell
 {
     /// call when user selected this cell
     override open func didSelected()